Minesweeper with methods
This C# project was created using Visual Studio 2013 Professional.  It's
a console application that allows the user to play Minesweeper.  It also
has some cheatcodes.  Looking back at it I believe the same result could
be achieved by using classes and it would require less code.  However my
current implementation works and it was rewarding to play a simple game
that I had created myself.  This code was based off an assignment I did
in my C# class in Summer 2015 BUT I added a lot of extra features and
code that wasn't in the assignment.